part1

+++ src/main/java/com/example/gestion_rh/service/CandidatService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.example.gestion_rh.model.Candidat;
-import com.example.gestion_rh.repository.Candidat;
+import com.example.gestion_rh.repository.CandidatRepository;
 
 @Service
 public class CandidatService {
@@ -17,9 +17,9 @@ public class CandidatService {
 
     public List<Candidat> getAll() { return repo.findAll(); }
 
-    public Candidat getById(Integer id) { return repo.findById(id).orElse(null); }
+    public Candidat getById(Long id) { return repo.findById(id).orElse(null); }
 
     public Candidat save(Candidat c) { return repo.save(c); }
 
-    public void delete(Integer id) { repo.deleteById(id); }
+    public void delete(Long id) { repo.deleteById(id); }
 }
